Move journey functionality to base class
