+++ flask-server/Dockerfile
@@ -1,7 +1,8 @@
 FROM python:3-alpine3.15
 WORKDIR /app
 
-COPY *.py requirements.txt /app/
+COPY *.py requirements.txt  models/ /app/
+
 RUN pip install -r requirements.txt
 
 EXPOSE 5000
