+++ flask-server/test_data.py
@@ -82,121 +82,3 @@ class Quest():
             self.status = QuestStatus.IN_PROGRESS 
         print(f"Quest Status == {self.status}\n\n")
         return {"success": "task completed."}
-class JourneyStatus(Enum): 
-    ACTIVE = 'Active'
-    INACTIVE = 'Inactive'
-
-class Journey():
-    def __init__(self, name: str, description: str):
-        self.name: str = name
-        self.description: str = description
-        self.quests: List[Quest]= [
-            Quest("This is a test quest name that's kinda long", "hello"),
-        ] 
-        self.status: JourneyStatus = JourneyStatus.INACTIVE
-        self.progress: int = 0
-    
-    def update_details(self):
-        pass
-
-    def add_quest(self, name, description):
-        self.quests.append(Quest(name, description))
-    
-    def remove_quest(self, questIdx):
-        self.quests.pop(questIdx)
-
-    def get_progress(self):
-        # count num complete tasks
-        all_tasks = sum( len(q.tasks) for q in self.quests)
-        completed_tasks = sum( sum(1 for task in quest.tasks if task.status == TaskStatus.COMPLETE) for quest in self.quests)
-        return int((completed_tasks/all_tasks)*100) if all_tasks > 0 else 0
-
-    def get_journey_details(self):
-        return {
-            'journeyName': self.name,
-            'journeyDetail': self.description,
-            'questsComplete': self.get_competed_quests(),
-            'totalQuests': len(self.quests),
-            'progress': self.get_progress(),
-            'questList': [q.name for q in self.quests]
-        }
-    
-    def get_competed_quests(self): 
-        complete = 0 
-        for q in self.quests: 
-            if q.status == QuestStatus.COMPLETE:
-                complete +=1
-        return complete
-    
-    def get_quests_preview(self):
-        return {'quests': [q.get_preview() for q in self.quests]}
-    
-    def get_quest_details(self, questIdx):
-        if questIdx < 0 or questIdx >= len(self.quests):
-            return {'error': "Index out of range"}
-        return self.quests[questIdx].get_preview()
-
-    def get_task_from_quest(self, questIdx, taskIdx):
-        if questIdx < 0 or questIdx >= len(self.quests):
-            return {'error': "Index out of range"}
-        return self.quests[questIdx].get_task_details(taskIdx)
-
-class GoalzillaData():
-    # Eventually may be good to put this in an external source...
-    def __init__(self, include_test_defaults: bool = False): 
-        self.goals = [] if not include_test_defaults else self.get_test_defaults()
-    
-    def get_goals_list(self):
-        return [goal.name for goal in self.goals]
-
-    def add_journey(self, name, description):
-        self.goals.append(Journey(name=name, description=description))
-
-    def get_journey_details(self, journeyIdx):
-        if journeyIdx < 0 or journeyIdx >= len(self.goals):
-            return None
-        return self.goals[journeyIdx].get_journey_details()
-    
-    def get_test_defaults(self):
-        # return []
-        return [
-            Journey(name = 'Test1', description = 'Hello1'),
-            Journey(name = 'Test2', description = 'Hello2'),
-            Journey(name = 'Test3', description = 'Hello3'), 
-        ]
-
-    def remove_journey(self, index):
-        self.goals.pop(index)
-
-    def get_quest_preview(self, journeyIdx):
-        return self.goals[journeyIdx].get_quests_preview()
-        
-    def add_quest(self, journeyIdx, name, description):
-        self.goals[journeyIdx].add_quest(name, description)
-
-    def remove_quest(self, journeyIdx, questIdx):
-        self.goals[journeyIdx].remove_quest(questIdx)
-
-    def get_quest_details(self, journeyIdx: int, questIdx: int):
-        return self.goals[journeyIdx].get_quest_details(questIdx)
-    
-    def get_task_details(self, journeyIdx, questIdx, taskIdx):
-         return self.goals[journeyIdx].get_task_from_quest(questIdx, taskIdx)
-    
-    def remove_task(self, journeyIdx, questIdx, taskIdx): 
-        journey = self.goals[journeyIdx]
-        quest: Quest = journey.quests[questIdx]
-        quest.tasks.pop(taskIdx)
-        quest.update_status()
-
-    def add_task(self, journeyIdx, questIdx, taskName):
-        journey = self.goals[journeyIdx]
-        quest: Quest = journey.quests[questIdx]
-        quest.tasks.append(Task(taskName))
-        quest.update_status()
-
-    def complete_task(self, journeyIdx, questIdx, taskIdx):
-        journey = self.goals[journeyIdx]
-        quest: Quest = journey.quests[questIdx]
-        print("in complete task main")
-        quest.complete_task(taskIdx)
